--- src/background.html
@@ -1,3 +1,0 @@
-<script type="module" src="./background/index.js"></script>
-
-<textarea id="clipboard-access"></textarea>
